requirements.txt
flask==3.0.0
werkzeug==3.0.1
openai-whisper==20231117
faster-whisper==1.0.0
torch==2.1.0
torchaudio==2.1.0
jiwer==3.0.3
librosa==0.10.1
soundfile==0.12.1
pydub==0.25.1
numpy==1.24.3
pandas==2.0.3
